organizing clean architecture pattern
