+++ app/src/main/java/org/hungrytessy/indycarsuperfan/data/remote/dto/Season.kt
@@ -1,4 +1,4 @@
-package org.hungrytessy.indycarsuperfan.data.models
+package org.hungrytessy.indycarsuperfan.data.remote.dto
 
 import java.util.TreeSet
