+++ app/src/main/java/org/hungrytessy/indycarsuperfan/data/remote/dto/Seasons.kt
@@ -1,0 +1,9 @@
+package org.hungrytessy.indycarsuperfan.data.remote.dto
+
+import com.google.gson.annotations.SerializedName
+import java.util.TreeSet
+
+data class Seasons(
+    @SerializedName("stages")
+    val stages: TreeSet<Season>
+)
