+++ app/src/main/java/org/hungrytessy/indycarsuperfan/data/remote/dto/Venue.kt
@@ -1,12 +1,13 @@
-package org.hungrytessy.indycarsuperfan.data.models
+package org.hungrytessy.indycarsuperfan.data.remote.dto
 
 import com.google.gson.annotations.SerializedName
-import org.hungrytessy.indycarsuperfan.data.IndyDataStore.drivers
 
 /**
  * only used to build a list of all venues
  */
-class Venues(@SerializedName("venues") val venues: HashMap<String, Venue> = HashMap()) {
+class Venues(
+    @SerializedName("venues") val venues: HashMap<String, Venue> = HashMap()
+) {
     constructor(vns: List<Venue>): this() {
         for (vn in vns) {
             venues[vn.id] = vn
@@ -28,20 +29,3 @@ data class Venue (
     @SerializedName("debut") val debut: Int?,
     @SerializedName("length") val length: Int?,
 )
-
-
-/*
-"curves_left": 4, "curves_right": 0, "url_official"
-  final String id;
-  final String? name;
-  final String? city;
-  final String? country;
-  final String? coordinates;
-  final String? countryCode;
-  final int? curvesRight;
-  final int? curvesLeft;
-  final int? debut;
-  final int? length;
-  final String? urlOfficial;
-  final String? timezone;
- */
