--- app/src/main/java/org/hungrytessy/indycarsuperfan/network/MainNetwork.kt
@@ -1,20 +1,0 @@
-package org.hungrytessy.indycarsuperfan.network
-
-import org.hungrytessy.indycarsuperfan.data.models.Drivers
-import org.hungrytessy.indycarsuperfan.data.models.Seasons
-import org.hungrytessy.indycarsuperfan.data.models.Venues
-import retrofit2.http.GET
-
-/**
- * Main network interface which will fetch a new welcome title for us
- */
-interface MainNetwork {
-    @GET(PATH_DRIVERS)
-    suspend fun getDrivers(): Drivers
-
-    @GET(PATH_SEASONS)
-    suspend fun getSeasons(): Seasons
-
-    @GET(PATH_VENUES)
-    suspend fun getVenues(): Venues
-}
