+++ app/src/main/java/org/hungrytessy/indycarsuperfan/ui/adapters/OnDriverClickListener.kt
@@ -1,6 +1,6 @@
 package org.hungrytessy.indycarsuperfan.ui.adapters
 
-import org.hungrytessy.indycarsuperfan.data.models.Driver
+import org.hungrytessy.indycarsuperfan.data.remote.dto.Driver
 
 interface OnDriverClickListener {
     fun onDriverClick(driver: Driver?)
