+++ app/src/main/java/org/hungrytessy/indycarsuperfan/ui/adapters/OnRaceClickListener.kt
@@ -1,7 +1,5 @@
 package org.hungrytessy.indycarsuperfan.ui.adapters
 
-import org.hungrytessy.indycarsuperfan.data.models.Stage
-
 interface OnRaceClickListener {
     fun onRaceClick(raceId: String)
 }
