+++ app/src/main/java/org/hungrytessy/indycarsuperfan/ui/results/singlerace/RaceStageResultViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import org.hungrytessy.indycarsuperfan.data.IndyDataStore
-import org.hungrytessy.indycarsuperfan.data.models.indy.SingleRaceStage
+import org.hungrytessy.indycarsuperfan.data.remote.dto.indy.SingleRaceStage
 
 class RaceStageResultViewModel : ViewModel() {
