+++ app/src/main/java/org/hungrytessy/indycarsuperfan/ui/schedule/singlerace/SingleRaceScheduleViewModel.kt
@@ -4,9 +4,9 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import org.hungrytessy.indycarsuperfan.data.IndyDataStore
-import org.hungrytessy.indycarsuperfan.data.models.Driver
-import org.hungrytessy.indycarsuperfan.data.models.Venue
-import org.hungrytessy.indycarsuperfan.data.models.indy.RaceWeekend
+import org.hungrytessy.indycarsuperfan.data.remote.dto.Driver
+import org.hungrytessy.indycarsuperfan.data.remote.dto.Venue
+import org.hungrytessy.indycarsuperfan.data.remote.dto.indy.RaceWeekend
 
 class SingleRaceScheduleViewModel : ViewModel() {
